PXE Provisioning fixtures and test using new infra (#9246)
* Start first test

* Add repo sync and HG creation

* Additions and comments

* rewrite test to fixtures

* Refactor test and fixtures

* Parametrized content fixture, changes to constants

* Creating host works, now let's use wrapanapi!

* Test edits

* Test enhancemets, docs addition

* Move fixtures to pytest_fixtures

* Use fixtures for LCE, CV and use different one for org.

* rename fixtures to reflect their scope in the name

* Use fixture for Subscription, parametrize the fixture using fixture_markers

* Fix typo, remove mock data

* Use power_control to power on the VM

* Use key-based authentication for the provisioned hosts

* Add `internal_capsule` property, address review comments

* Increase timeout, allow the host to provision

* Move the test logic from cli to api directory, rename test, change the docstring.

+++ robottelo/hosts.py
@@ -137,6 +137,7 @@ class ContentHost(Host, ContentHostMixins):
             # key file based authentication
             kwargs.update({'key_filename': auth})
         self._satellite = kwargs.get('satellite')
+        self.blank = kwargs.get('blank', False)
         super().__init__(hostname=hostname, **kwargs)
 
     @property
@@ -195,13 +196,15 @@ class ContentHost(Host, ContentHostMixins):
         return Version(version=version_string)
 
     def setup(self):
-        self.remove_katello_ca()
-        self.execute('subscription-manager clean')
+        if not self.blank:
+            self.remove_katello_ca()
+            self.execute('subscription-manager clean')
 
     def teardown(self):
-        if self.nailgun_host:
-            self.nailgun_host.delete()
-        self.unregister()
+        if not self.blank:
+            if self.nailgun_host:
+                self.nailgun_host.delete()
+            self.unregister()
         # Strip most unnecessary attributes from our instance for checkin
         keep_keys = set(self.to_dict()) | {'release', '_prov_inst'}
         self.__dict__ = {k: v for k, v in self.__dict__.items() if k in keep_keys}
@@ -1355,6 +1358,11 @@ class Satellite(Capsule, SatelliteMixins):
                         # not everything has an mro method, we don't care about them
                         pass
         return self._cli
+
+    @property
+    def internal_capsule(self):
+        capsule_list = self.api.SmartProxy().search(query={'search': f'name={self.hostname}'})
+        return None if not capsule_list else capsule_list[0]
 
     def ui_session(self, testname=None, user=None, password=None, url=None, login=True):
         """Initialize an airgun Session object and store it as self.ui_session"""
